fixed setup and test issues
