+++ backend/.gitignore
@@ -1,1 +1,3 @@
 node_modules
+dist
+.nyc_output
