+++ backend/app.ts
@@ -1,12 +1,6 @@
-import express from "express";
-import cors from "cors";
-const app = express();
+import app from "./src/server";
 const port = 3000;
 
-app.use(express.text());
-
-app.use(cors());
-
-app.listen(port);
-
-export default app;
+// app.listen(port, () => {
+//   console.log("RUNNING SERVER");
+// });
