+++ backend/package.json
@@ -1,30 +1,31 @@
 {
-  "dependencies": {
-    "@mozilla/readability": "^0.4.2",
-    "@nlpjs/lang-en": "^4.23.4",
-    "@types/isomorphic-fetch": "^0.0.36",
-    "cors": "^2.8.5",
-    "express": "^4.18.2",
-    "jsdom": "^20.0.2"
-  },
-  "type": "module",
-  "scripts": {
-    "test": "mocha --require ts-node/register src/test/* --reporter spec"
-  },
-  "devDependencies": {
-    "@types/chai": "^4.3.4",
-    "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.14",
-    "@types/mocha": "^10.0.0",
-    "@types/node": "^18.11.9",
-    "babel-cli": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "chai": "^4.3.7",
-    "mocha": "^10.1.0",
-    "superagent": "^8.0.3",
-    "supertest": "^6.3.1",
-    "ts-jest": "^29.0.3",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.8.4"
-  }
+    "dependencies": {
+        "@mozilla/readability": "^0.4.2",
+        "@nlpjs/lang-en": "^4.23.4",
+        "@types/isomorphic-fetch": "^0.0.36",
+        "cors": "^2.8.5",
+        "express": "^4.18.2",
+        "jsdom": "^20.0.2"
+    },
+    "scripts": {
+        "start": " tsc && node dist/app.js",
+        "test": "nyc mocha --require ts-node/register 'test/**/*.spec.ts'"
+    },
+    "devDependencies": {
+        "@types/chai": "^4.3.4",
+        "@types/cors": "^2.8.12",
+        "@types/express": "^4.17.14",
+        "@types/mocha": "^10.0.0",
+        "@types/node": "^18.11.9",
+        "babel-cli": "^6.26.0",
+        "babel-preset-env": "^1.7.0",
+        "chai": "^4.3.7",
+        "mocha": "^10.1.0",
+        "nyc": "^15.1.0",
+        "superagent": "^8.0.3",
+        "supertest": "^6.3.1",
+        "ts-jest": "^29.0.3",
+        "ts-node": "^10.9.1",
+        "typescript": "^4.8.4"
+    }
 }
