+++ backend/test/keywords.spec.ts
@@ -1,0 +1,29 @@
+import { expect } from "chai";
+
+import { getKeywordsFromText } from "../src/keyword";
+
+describe("Keyword Extraction", function () {
+  describe("Extract keyword from text", function () {
+    it("extract words that occur most frequent", function () {
+      const testString = "Hello, my name is X, what is your name?";
+
+      expect(getKeywordsFromText(testString, 1)[0]).to.equal("name");
+    });
+
+    it("find expected number of keyword", function () {
+      const testString =
+        "I am trying to test my firebase security rules, using Typescript and Mocha, but I can't even run my test";
+      expect(getKeywordsFromText(testString, 3).length).to.equal(3);
+    });
+
+    it("not influenced by special character", function () {
+      const testString = "I feel so bored. . . . ., too bored!";
+      expect(getKeywordsFromText(testString, 1)[0]).to.equal("bored");
+    });
+
+    it("word in the same semantic meaning but different form will be considered one", function () {
+      const testString = "These classes are not following class convention, class, classes";
+      expect(getKeywordsFromText(testString, 1)[1]).to.not.equal("classes");
+    });
+  });
+});
