+++ backend/tsconfig.json
@@ -1,20 +1,22 @@
 {
   "compilerOptions": {
+    "noImplicitAny": false,
+    // "noImplicitReturns": true,
+    // "noImplicitThis": true,
+    // "noUnusedLocals": true,
+    // "noUnusedParameters": false,
+    // "preserveConstEnums": true,
+    // "strictFunctionTypes": true,
+    // "strictNullChecks": true,
+    "target": "es2016",
+    "moduleResolution": "node",
 
-    /* Modules */
-    "module": "ES2020",                                /* Specify what module code is generated. */
-    "typeRoots": ["./node_modules/@types"],              /* Specify multiple folders that act like './node_modules/@types'. */
-    "outDir": "./",                                      /* Specify an output folder for all emitted files. */
-    "allowSyntheticDefaultImports": true,                /* Allow 'import x from y' when a module doesn't have a default export. */
-    "esModuleInterop": true,                             /* Emit additional JavaScript to ease support for importing CommonJS modules. This enables 'allowSyntheticDefaultImports' for type compatibility. */
-    "forceConsistentCasingInFileNames": true,            /* Ensure that casing is correct in imports. */
-    "strict": true,                                      /* Enable all strict type-checking options. */
-    "skipLibCheck": true,                                 /* Skip type checking all .d.ts files. */
+    "outDir": "dist",
+    "esModuleInterop": true,
+    // "forceConsistentCasingInFileNames": true,
+    // "strict": true,
+    "skipLibCheck": true
   },
-  "include": [
-    "src/**/*.ts"
-  ],
-  "ts-node": {
-    "esm": true
-  }
+  "include": ["./src/**/*.ts"],
+  "exclude": ["./tests/", "./node_modules/", "./dist/"]
 }
